cf4k8s-gs.md fix a yaml key and a CLI quote syntax (#816)
New users may have had trouble installing CF following this guide because of a recent regression. The name of the yaml key for the users dockerhub acct was changed from repository to repository_prefix for clarity on how it is used. Unfortunately this change has not made it back into this guide until now. In addition, the ytt config command at the end of the section Generate the yaml used to deploy CF for k8s needs to be run to complete setup but it is not quoted in the same style as all of the other copy paste items that are intended fo the user to copy paste into a command line.